feat: Allow a Kind to have a adhoc case
So the question is what if someone wants programmatic mixed with
non-programmatic?  One option is to make the `Kind` an `Option<>` and
have a separate field for adhoc messages. I was concerned at how that'd
turn out.

I instead came up with this idea.  What if the user's `Kind` had a
variant that was `doc(hidden)` and stored a `&'static str`?  That would
allow theu ser to control how strictly typed their API is.